Moved batched update helpers to BatchedUpdateUtil
